Fix: RSA decrupt

+++ src/player.h
@@ -414,7 +414,7 @@ class Player final : public Creature, public Cylinder
 			operatingSystem = clientos;
 		}
 
-		uint32_t getProtocolVersion() const {
+		uint16_t getProtocolVersion() const {
 			if (!client) {
 				return 0;
 			}
